requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.11

# 截图功能（可选，仅 quick_image 模式需要）
selenium>=4.15.0
webdriver-manager>=4.0.0
